pyproject.toml
[project]
name = "todoist-mcp-server"
version = "0.1.0"
description = "MCP server for Todoist integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "todoist-api-python>=3.1.0",
]
